+++ src/vtex/getProductSuggestions/index.ts
@@ -1,0 +1,3 @@
+import getProductSuggestions from './getProductSuggestions';
+
+export default getProductSuggestions;