+++ src/vtex/index.ts
@@ -1,4 +1,5 @@
 import MasterData from './MasterData';
 import getProductById from './getProductById';
+import getProductSuggestions from './getProductSuggestions';
 
-export { MasterData, getProductById };
+export { MasterData, getProductById, getProductSuggestions };